requirements.txt
claude-agent-sdk>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-docx>=1.0.0
weasyprint>=60.0
markdown>=3.5.0
requests>=2.31.0
python-dotenv>=1.0.0
